requirements.txt
pygame
numpy
timeit
PyOpenGL
PyOpenGL-accelerate
pillow
